requirements.txt
tkinter
requests
datetime
matplotlib
mysql-connector-python